refactor(input): consolidate provider overlay into 5 cohesive files
The provider selector was split across 8 non-test files, so a single user
action spanned many of them. Regroup into 5 files by responsibility with no
logic change: definitions+state+dispatch (on_provider.go), navigation+selection
(on_provider_nav.go), credentials+connect (on_provider_auth.go), data loading
(on_provider_data.go), and rendering (on_provider_view.go).
